mainly minor fixes for readability
The additional "poorly written or malicious software" is important- in theory M-mode can lift PMP restrictions, but only if such software exists to be run during the boot process. In this case the software must be malicious (someone is attacking the system and injecting code) or poorly written (existing code allows such an attack).